% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot a clustering graph from Laplacian matrix

Plots a clustering graph based on a given Laplacian matrix and the ground-truth labels for the clusters.}
\usage{
plot_graph(Laplacian, true_labels, names)
}
\arguments{
\item{Laplacian}{r by r Laplacian matrix.}

\item{true_labels}{the numeric vector of true labels (of length r)}

\item{names}{the vector of names of the elements, e.g., stocks (of length r)}
}
\value{
The network object:
\item{\code{graph_net}}{the graph network object that can be used for plotting}
\item{\code{metric}}{the results of the clustering measures}
}
\description{
Plot a clustering graph from Laplacian matrix

Plots a clustering graph based on a given Laplacian matrix and the ground-truth labels for the clusters.
}
